german_strings.xml
Here is the german translation of your service. Thanks for the chance to get a part of it. Hope the umlauts are not a problem in UTF-8, otherwise i will change it. Is it possible to see is in a beta? So i can check wheather i translated it correctly.
Have a nice day Tim