Fixed issue #125 - puh notifications do not always work
Problem is that the code which sends the request to the GCM is inside an AsyncTask and since Froyo tasks are run serially, so if another task starts before it and doesn't finish it doesn't run.
The fix is to use the new API of executeOnExecutor which runs the task in parallel to other tasks.