.nomedia file in cache
On some phones (e.g. HTC M8) gallery apps (e.g. HTC Gallery) show the Telegram cache folder as an album. 
Even though this is not the fault of telegram app I believe for the UX it would be better if the app created a .nomedia file, so the (not so perfect) gallery apps skip scanning the cache folder.