Use rounded to even resultWidth & resultHeight
Had an issue (reproduced in your App) when transcoding from a 620x714 video input to a scaled down 555x640 output. The encoder was throwing during configuration. After some research this is apparently caused by odd output sizes (in the sample 555). Just by rounding to nearest even number this issue gets fixed.

I added the rounding code to the VideoEditorActivity.java since is where those values are being computed and the rounding happens, but this could be done/checked later on during transcoding in MediaController.